==> CA1/03 Data Pre-Processing.ipynb <==
# Data Pre-Processing

Perform data cleaning on the consolidated dataset, including the following:
<ol>
    <li>Convert all text to Lower Case</li>
    <li>Remove special breakline characters</li>
    <li>Remove Numbers</li>
    <li>Lemmatization of text</li>
    <li>Remove Stopwords</li>
    <li>Category Encoding</li>
</ol>

The output is a cleaned dataset.

import pandas as pd
import pickle
import re

import nltk
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
nltk.download('stopwords')

%matplotlib inline

path_df = "./Pickles/sampled_articles_raw.pickle"

with open(path_df, 'rb') as data:
    articles = pickle.load(data)

## 1. Convert All text to Lower Case

articles['article']=articles['article'].str.lower()

articles['article'].iloc[0]

## 2. Remove special breakline characters and Punctuation

articles['article']=articles['article'].str.replace("\n", " ")
articles['article']=articles['article'].str.replace(r'[^\w\s]+', '')

articles['article'].iloc[0]

## 3. Remove Numbers

articles['article'] = articles['article'].str.replace('\d+', '')

articles['article'].iloc[0]

## 4. Lemmatization of text

lemmatizer = WordNetLemmatizer() 

def lemmatize_text(raw_text):
    raw_text_words = raw_text.split(" ")
    
    lemmatized_text_list = []
    
    for word in raw_text_words:
        lemmatized_text_list.append(lemmatizer.lemmatize(word, pos="v"))
        
    lemmatized_text = " ".join(lemmatized_text_list)
    
    return lemmatized_text

articles['article']=articles['article'].apply(lemmatize_text)

articles['article'].iloc[0]

## 5. Remove Stopwords

stop_words = list(stopwords.words('english'))

def remove_stopwords(raw_text):
    
    raw_text_words = raw_text.split(" ")
    
    stopwords_removed_text_list = []
    
    for word in raw_text_words:
        if word.lower() not in stop_words:
            stopwords_removed_text_list.append(word)
        
    stopwords_removed_text = " ".join(stopwords_removed_text_list)
    
    return stopwords_removed_text

articles['article']=articles['article'].apply(remove_stopwords)

articles['article'].iloc[0]

## 6. Category Encoding

articles['category'].unique()

category_mapping = {
    'Singapore': 1,
    'Sports': 2,
    'Lifestyle': 3,
    'World': 4,
    'Business': 5,
    'Technology': 6
}

articles['category_code']=articles['category']
processed_articles = articles.replace({'category_code':category_mapping})
processed_articles=processed_articles.reset_index(drop=True)
processed_articles.head()

#Export to Serialized Object
with open('Pickles/all_articles_processed.pickle', 'wb') as output:
    pickle.dump(processed_articles, output)
